# packages/api-types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mini Hedge Fund Desk",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/orders": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Create Order",
        "operationId": "create_order_api_v1_orders_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "List Orders",
        "operationId": "list_orders_api_v1_orders_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderSummary"
                  },
                  "title": "Response List Orders Api V1 Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get Order",
        "operationId": "get_order_api_v1_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/fills": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get Fills",
        "operationId": "get_fills_api_v1_orders__order_id__fills_get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FillDetail"
                  },
                  "title": "Response Get Fills Api V1 Orders  Order Id  Fills Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/algo": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Create Algo Order",
        "operationId": "create_algo_order_api_v1_orders_algo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlgoOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/children": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get Children",
        "operationId": "get_children_api_v1_orders__order_id__children_get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderSummary"
                  },
                  "title": "Response Get Children Api V1 Orders  Order Id  Children Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}/cancel": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Cancel Order",
        "operationId": "cancel_order_api_v1_orders__order_id__cancel_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/allocations": {
      "post": {
        "tags": [
          "allocations"
        ],
        "summary": "Create Block Allocation",
        "operationId": "create_block_allocation_api_v1_allocations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlockAllocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockAllocationSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/allocations/{allocation_id}": {
      "get": {
        "tags": [
          "allocations"
        ],
        "summary": "Get Allocation",
        "operationId": "get_allocation_api_v1_allocations__allocation_id__get",
        "parameters": [
          {
            "name": "allocation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Allocation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockAllocationSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/allocations/{allocation_id}/cancel": {
      "post": {
        "tags": [
          "allocations"
        ],
        "summary": "Cancel Allocation",
        "operationId": "cancel_allocation_api_v1_allocations__allocation_id__cancel_post",
        "parameters": [
          {
            "name": "allocation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Allocation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockAllocationSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlgoParams": {
        "properties": {
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds",
            "default": 7200
          },
          "num_slices": {
            "type": "integer",
            "title": "Num Slices",
            "default": 200
          },
          "visible_quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible Quantity"
          },
          "volume_profile": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "string",
                      "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Profile"
          }
        },
        "type": "object",
        "title": "AlgoParams"
      },
      "AlgoType": {
        "type": "string",
        "enum": [
          "twap",
          "vwap",
          "iceberg"
        ],
        "title": "AlgoType"
      },
      "AllocationLegRequest": {
        "properties": {
          "fund_slug": {
            "type": "string",
            "title": "Fund Slug"
          },
          "portfolio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Portfolio Id"
          },
          "target_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Target Pct"
          }
        },
        "type": "object",
        "required": [
          "fund_slug",
          "portfolio_id",
          "target_pct"
        ],
        "title": "AllocationLegRequest"
      },
...
      "BlockAllocationSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "side": {
            "type": "string",
            "title": "Side"
          },
          "total_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Quantity"
          },
          "filled_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Filled Quantity"
          },
          "avg_fill_price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Fill Price"
          },
          "state": {
            "$ref": "#/components/schemas/AllocationState"
          },
          "algo_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlgoType"
              },
              {
                "type": "null"
              }
            ]
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/AllocationLegSummary"
            },
            "type": "array",
            "title": "Legs"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "instrument_id",
          "side",
          "total_quantity",
          "filled_quantity",
          "avg_fill_price",
          "state",
          "legs",
          "created_by",
          "created_at"
        ],
        "title": "BlockAllocationSummary"
      },
...
      "CreateAlgoOrderRequest": {
        "properties": {
          "portfolio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Portfolio Id"
          },
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "side": {
            "$ref": "#/components/schemas/OrderSide"
          },
          "order_type": {
            "$ref": "#/components/schemas/OrderType",
            "default": "limit"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Quantity"
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "time_in_force": {
            "$ref": "#/components/schemas/TimeInForce",
            "default": "day"
          },
          "algo_type": {
            "$ref": "#/components/schemas/AlgoType"
          },
          "algo_params": {
            "$ref": "#/components/schemas/AlgoParams",
            "default": {
              "duration_seconds": 7200,
              "num_slices": 200
            }
          }
        },
        "type": "object",
        "required": [
          "portfolio_id",
          "instrument_id",
          "side",
          "quantity",
          "algo_type"
        ],
        "title": "CreateAlgoOrderRequest"
      },
      "CreateBlockAllocationRequest": {
        "properties": {
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "side": {
            "$ref": "#/components/schemas/OrderSide"
          },
          "total_quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Total Quantity"
          },
          "order_type": {
            "$ref": "#/components/schemas/OrderType",
            "default": "market"
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "algo_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlgoType"
              },
              {
                "type": "null"
              }
            ]
          },
          "algo_params": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlgoParams"
              },
              {
                "type": "null"
              }
            ]
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/AllocationLegRequest"
            },
            "type": "array",
            "title": "Legs"
          }
        },
        "type": "object",
        "required": [
          "instrument_id",
          "side",
          "total_quantity",
          "legs"
        ],
        "title": "CreateBlockAllocationRequest"
      },
...
      "CreateOrderRequest": {
        "properties": {
          "portfolio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Portfolio Id"
          },
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "side": {
            "$ref": "#/components/schemas/OrderSide"
          },
          "order_type": {
            "$ref": "#/components/schemas/OrderType",
            "default": "market"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Quantity"
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "time_in_force": {
            "$ref": "#/components/schemas/TimeInForce",
            "default": "day"
          }
        },
        "type": "object",
        "required": [
          "portfolio_id",
          "instrument_id",
          "side",
          "quantity"
        ],
        "title": "CreateOrderRequest"
      },
...
      "FillDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "order_id": {
            "type": "string",
            "format": "uuid",
            "title": "Order Id"
          },
          "quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "filled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Filled At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "quantity",
          "price",
          "filled_at"
        ],
        "title": "FillDetail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderSide": {
        "type": "string",
        "enum": [
          "buy",
          "sell"
        ],
        "title": "OrderSide"
      },
      "OrderState": {
        "type": "string",
        "enum": [
          "draft",
          "pending_compliance",
          "approved",
          "rejected",
          "sent",
          "working",
          "partially_filled",
          "filled",
          "cancelled"
        ],
        "title": "OrderState"
      },
      "OrderSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "portfolio_id": {
            "type": "string",
            "format": "uuid",
            "title": "Portfolio Id"
          },
          "instrument_id": {
            "type": "string",
            "title": "Instrument Id"
          },
          "side": {
            "$ref": "#/components/schemas/OrderSide"
          },
          "order_type": {
            "$ref": "#/components/schemas/OrderType"
          },
          "quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity"
          },
          "filled_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Filled Quantity"
          },
          "limit_price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Price"
          },
          "avg_fill_price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Fill Price"
          },
          "state": {
            "$ref": "#/components/schemas/OrderState"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "compliance_results": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliance Results"
          },
          "time_in_force": {
            "$ref": "#/components/schemas/TimeInForce"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "algo_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlgoType"
              },
              {
                "type": "null"
              }
            ]
          },
          "algo_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Algo Params"
          },
          "is_parent": {
            "type": "boolean",
            "title": "Is Parent",
            "default": false
          },
          "parent_order_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Order Id"
          },
          "children_count": {
            "type": "integer",
            "title": "Children Count",
            "default": 0
          },
          "children_filled": {
            "type": "integer",
            "title": "Children Filled",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "portfolio_id",
          "instrument_id",
          "side",
          "order_type",
          "quantity",
          "filled_quantity",
          "limit_price",
          "avg_fill_price",
          "state",
          "rejection_reason",
          "compliance_results",
          "time_in_force",
          "created_at",
          "updated_at"
        ],
        "title": "OrderSummary"
      },
      "OrderType": {
        "type": "string",
        "enum": [
          "market",
          "limit"
        ],
        "title": "OrderType"
      },
...
      "TimeInForce": {
        "type": "string",
        "enum": [
          "day",
          "gtc",
          "ioc"
        ],
        "title": "TimeInForce"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}